% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_objects.R
\name{Account}
\alias{Account}
\title{Google Mirror API Objects 
Interacts with Glass users via the timeline.}
\usage{
Account(authTokens = NULL, features = NULL, password = NULL,
  userData = NULL)
}
\arguments{
\item{authTokens}{No description}

\item{features}{No description}

\item{password}{No description}

\item{userData}{No description}
}
\value{
Account object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:24:14
filename: /Users/mark/dev/R/autoGoogleAPI/googlemirrorv1.auto/R/mirror_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Account Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents an account passed into the Account Manager on Glass.
}
\seealso{
Other Account functions: \code{\link{accounts.insert}}
}

