% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_functions.R
\name{accounts.insert}
\alias{accounts.insert}
\title{Inserts a new account for a user}
\usage{
accounts.insert(Account, userToken, accountType, accountName)
}
\arguments{
\item{Account}{The \link{Account} object to pass to this method}

\item{userToken}{The ID for the user}

\item{accountType}{Account type to be passed to Android Account Manager}

\item{accountName}{The name of the account to be passed to the Android Account Manager}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/glass}{Google Documentation}

Other Account functions: \code{\link{Account}}
}
