% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_objects.R
\name{Attachment}
\alias{Attachment}
\title{Attachment Object}
\usage{
Attachment(contentType = NULL, contentUrl = NULL, id = NULL,
  isProcessingContent = NULL)
}
\arguments{
\item{contentType}{The MIME type of the attachment}

\item{contentUrl}{The URL for the content}

\item{id}{The ID of the attachment}

\item{isProcessingContent}{Indicates that the contentUrl is not available because the attachment content is still being processed}
}
\value{
Attachment object
}
\description{
Attachment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents media content, such as a photo, that can be attached to a timeline item.
}
