% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_objects.R
\name{AttachmentsListResponse}
\alias{AttachmentsListResponse}
\title{AttachmentsListResponse Object}
\usage{
AttachmentsListResponse(items = NULL)
}
\arguments{
\item{items}{The list of attachments}
}
\value{
AttachmentsListResponse object
}
\description{
AttachmentsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of Attachments. This is the response from the server to GET requests on the attachments collection.
}
