% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_objects.R
\name{AuthToken}
\alias{AuthToken}
\title{AuthToken Object}
\usage{
AuthToken(authToken = NULL, type = NULL)
}
\arguments{
\item{authToken}{No description}

\item{type}{No description}
}
\value{
AuthToken object
}
\description{
AuthToken Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
