% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_objects.R
\name{Command}
\alias{Command}
\title{Command Object}
\usage{
Command(type = NULL)
}
\arguments{
\item{type}{The type of operation this command corresponds to}
}
\value{
Command object
}
\description{
Command Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single menu command that is part of a Contact.
}
