% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_objects.R
\name{Contact}
\alias{Contact}
\title{Contact Object}
\usage{
Contact(acceptCommands = NULL, acceptTypes = NULL, displayName = NULL,
  id = NULL, imageUrls = NULL, phoneNumber = NULL, priority = NULL,
  sharingFeatures = NULL, source = NULL, speakableName = NULL,
  type = NULL)
}
\arguments{
\item{acceptCommands}{A list of voice menu commands that a contact can handle}

\item{acceptTypes}{A list of MIME types that a contact supports}

\item{displayName}{The name to display for this contact}

\item{id}{An ID for this contact}

\item{imageUrls}{Set of image URLs to display for a contact}

\item{phoneNumber}{Primary phone number for the contact}

\item{priority}{Priority for the contact to determine ordering in a list of contacts}

\item{sharingFeatures}{A list of sharing features that a contact can handle}

\item{source}{The ID of the application that created this contact}

\item{speakableName}{Name of this contact as it should be pronounced}

\item{type}{The type for this contact}
}
\value{
Contact object
}
\description{
Contact Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person or group that can be used as a creator or a contact.
}
\seealso{
Other Contact functions: \code{\link{contacts.insert}},
  \code{\link{contacts.patch}},
  \code{\link{contacts.update}}
}

