% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_objects.R
\name{ContactsListResponse}
\alias{ContactsListResponse}
\title{ContactsListResponse Object}
\usage{
ContactsListResponse(items = NULL)
}
\arguments{
\item{items}{Contact list}
}
\value{
ContactsListResponse object
}
\description{
ContactsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of Contacts representing contacts. This is the response from the server to GET requests on the contacts collection.
}

