% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_objects.R
\name{Location}
\alias{Location}
\title{Location Object}
\usage{
Location(accuracy = NULL, address = NULL, displayName = NULL, id = NULL,
  latitude = NULL, longitude = NULL, timestamp = NULL)
}
\arguments{
\item{accuracy}{The accuracy of the location fix in meters}

\item{address}{The full address of the location}

\item{displayName}{The name to be displayed}

\item{id}{The ID of the location}

\item{latitude}{The latitude, in degrees}

\item{longitude}{The longitude, in degrees}

\item{timestamp}{The time at which this location was captured, formatted according to RFC 3339}
}
\value{
Location object
}
\description{
Location Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A geographic location that can be associated with a timeline item.
}

