% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_objects.R
\name{LocationsListResponse}
\alias{LocationsListResponse}
\title{LocationsListResponse Object}
\usage{
LocationsListResponse(items = NULL)
}
\arguments{
\item{items}{The list of locations}
}
\value{
LocationsListResponse object
}
\description{
LocationsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of Locations. This is the response from the server to GET requests on the locations collection.
}

