% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_objects.R
\name{MenuItem}
\alias{MenuItem}
\title{MenuItem Object}
\usage{
MenuItem(action = NULL, contextual_command = NULL, id = NULL,
  payload = NULL, removeWhenSelected = NULL, values = NULL)
}
\arguments{
\item{action}{Controls the behavior when the user picks the menu option}

\item{contextual_command}{The ContextualMenus}

\item{id}{The ID for this menu item}

\item{payload}{A generic payload whose meaning changes depending on this MenuItem's action}

\item{removeWhenSelected}{If set to true on a CUSTOM menu item, that item will be removed from the menu after it is selected}

\item{values}{For CUSTOM items, a list of values controlling the appearance of the menu item in each of its states}
}
\value{
MenuItem object
}
\description{
MenuItem Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A custom menu item that can be presented to the user by a timeline item.
}
