% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_objects.R
\name{MenuValue}
\alias{MenuValue}
\title{MenuValue Object}
\usage{
MenuValue(displayName = NULL, iconUrl = NULL, state = NULL)
}
\arguments{
\item{displayName}{The name to display for the menu item}

\item{iconUrl}{URL of an icon to display with the menu item}

\item{state}{The state that this value applies to}
}
\value{
MenuValue object
}
\description{
MenuValue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single value that is part of a MenuItem.
}
