% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_objects.R
\name{Notification}
\alias{Notification}
\title{Notification Object}
\usage{
Notification(collection = NULL, itemId = NULL, operation = NULL,
  userActions = NULL, userToken = NULL, verifyToken = NULL)
}
\arguments{
\item{collection}{The collection that generated the notification}

\item{itemId}{The ID of the item that generated the notification}

\item{operation}{The type of operation that generated the notification}

\item{userActions}{A list of actions taken by the user that triggered the notification}

\item{userToken}{The user token provided by the service when it subscribed for notifications}

\item{verifyToken}{The secret verify token provided by the service when it subscribed for notifications}
}
\value{
Notification object
}
\description{
Notification Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A notification delivered by the API.
}

