% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_objects.R
\name{NotificationConfig}
\alias{NotificationConfig}
\title{NotificationConfig Object}
\usage{
NotificationConfig(deliveryTime = NULL, level = NULL)
}
\arguments{
\item{deliveryTime}{The time at which the notification should be delivered}

\item{level}{Describes how important the notification is}
}
\value{
NotificationConfig object
}
\description{
NotificationConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Controls how notifications for a timeline item are presented to the user.
}
