% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_objects.R
\name{Setting}
\alias{Setting}
\title{Setting Object}
\usage{
Setting(id = NULL, value = NULL)
}
\arguments{
\item{id}{The setting's ID}

\item{value}{The setting value, as a string}
}
\value{
Setting object
}
\description{
Setting Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A setting for Glass.
}
