% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_functions.R
\name{settings.get}
\alias{settings.get}
\title{Gets a single setting by ID.}
\usage{
settings.get(id)
}
\arguments{
\item{id}{The ID of the setting}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/glass.timeline
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/glass.timeline)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/glass}{Google Documentation}
}

