% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_objects.R
\name{Subscription}
\alias{Subscription}
\title{Subscription Object}
\usage{
Subscription(callbackUrl = NULL, collection = NULL, id = NULL,
  notification = NULL, operation = NULL, updated = NULL,
  userToken = NULL, verifyToken = NULL)
}
\arguments{
\item{callbackUrl}{The URL where notifications should be delivered (must start with https://)}

\item{collection}{The collection to subscribe to}

\item{id}{The ID of the subscription}

\item{notification}{Container object for notifications}

\item{operation}{A list of operations that should be subscribed to}

\item{updated}{The time at which this subscription was last modified, formatted according to RFC 3339}

\item{userToken}{An opaque token sent to the subscriber in notifications so that it can determine the ID of the user}

\item{verifyToken}{A secret token sent to the subscriber in notifications so that it can verify that the notification was generated by Google}
}
\value{
Subscription object
}
\description{
Subscription Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A subscription to events on a collection.
}
\seealso{
Other Subscription functions: \code{\link{subscriptions.insert}},
  \code{\link{subscriptions.update}}
}

