% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_functions.R
\name{subscriptions.insert}
\alias{subscriptions.insert}
\title{Creates a new subscription.}
\usage{
subscriptions.insert(Subscription)
}
\arguments{
\item{Subscription}{The \link{Subscription} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/glass.timeline
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/glass.timeline)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/glass}{Google Documentation}

Other Subscription functions: \code{\link{Subscription}},
  \code{\link{subscriptions.update}}
}
