% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_functions.R
\name{subscriptions.list}
\alias{subscriptions.list}
\title{Retrieves a list of subscriptions for the authenticated user and service.}
\usage{
subscriptions.list()
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/glass.timeline
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/glass.timeline)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details. 

#' @importFrom googleAuthR gar_api_generator
}
\seealso{
\href{https://developers.google.com/glass}{Google Documentation}
}

