% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_objects.R
\name{SubscriptionsListResponse}
\alias{SubscriptionsListResponse}
\title{SubscriptionsListResponse Object}
\usage{
SubscriptionsListResponse(items = NULL)
}
\arguments{
\item{items}{The list of subscriptions}
}
\value{
SubscriptionsListResponse object
}
\description{
SubscriptionsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of Subscriptions. This is the response from the server to GET requests on the subscription collection.
}

