% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_functions.R
\name{timeline.list}
\alias{timeline.list}
\title{Retrieves a list of timeline items for the authenticated user.}
\usage{
timeline.list(bundleId = NULL, includeDeleted = NULL, maxResults = NULL,
  orderBy = NULL, pageToken = NULL, pinnedOnly = NULL,
  sourceItemId = NULL)
}
\arguments{
\item{bundleId}{If provided, only items with the given bundleId will be returned}

\item{includeDeleted}{If true, tombstone records for deleted items will be returned}

\item{maxResults}{The maximum number of items to include in the response, used for paging}

\item{orderBy}{Controls the order in which timeline items are returned}

\item{pageToken}{Token for the page of results to return}

\item{pinnedOnly}{If true, only pinned items will be returned}

\item{sourceItemId}{If provided, only items with the given sourceItemId will be returned}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/glass.location
\item https://www.googleapis.com/auth/glass.timeline
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/glass.location, https://www.googleapis.com/auth/glass.timeline)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/glass}{Google Documentation}
}
