% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_functions.R
\name{timeline.update}
\alias{timeline.update}
\title{Updates a timeline item in place.}
\usage{
timeline.update(TimelineItem, id)
}
\arguments{
\item{TimelineItem}{The \link{TimelineItem} object to pass to this method}

\item{id}{The ID of the timeline item}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/glass.location
\item https://www.googleapis.com/auth/glass.timeline
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/glass.location, https://www.googleapis.com/auth/glass.timeline)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/glass}{Google Documentation}

Other TimelineItem functions: \code{\link{TimelineItem}},
  \code{\link{timeline.insert}},
  \code{\link{timeline.patch}}
}
