% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_objects.R
\name{TimelineItem}
\alias{TimelineItem}
\title{TimelineItem Object}
\usage{
TimelineItem(attachments = NULL, bundleId = NULL, canonicalUrl = NULL,
  created = NULL, creator = NULL, displayTime = NULL, etag = NULL,
  html = NULL, id = NULL, inReplyTo = NULL, isBundleCover = NULL,
  isDeleted = NULL, isPinned = NULL, location = NULL, menuItems = NULL,
  notification = NULL, pinScore = NULL, recipients = NULL,
  selfLink = NULL, sourceItemId = NULL, speakableText = NULL,
  speakableType = NULL, text = NULL, title = NULL, updated = NULL)
}
\arguments{
\item{attachments}{A list of media attachments associated with this item}

\item{bundleId}{The bundle ID for this item}

\item{canonicalUrl}{A canonical URL pointing to the canonical/high quality version of the data represented by the timeline item}

\item{created}{The time at which this item was created, formatted according to RFC 3339}

\item{creator}{The user or group that created this item}

\item{displayTime}{The time that should be displayed when this item is viewed in the timeline, formatted according to RFC 3339}

\item{etag}{ETag for this item}

\item{html}{HTML content for this item}

\item{id}{The ID of the timeline item}

\item{inReplyTo}{If this item was generated as a reply to another item, this field will be set to the ID of the item being replied to}

\item{isBundleCover}{Whether this item is a bundle cover}

\item{isDeleted}{When true, indicates this item is deleted, and only the ID property is set}

\item{isPinned}{When true, indicates this item is pinned, which means it's grouped alongside 'active' items like navigation and hangouts, on the opposite side of the home screen from historical (non-pinned) timeline items}

\item{location}{The geographic location associated with this item}

\item{menuItems}{A list of menu items that will be presented to the user when this item is selected in the timeline}

\item{notification}{Controls how notifications for this item are presented on the device}

\item{pinScore}{For pinned items, this determines the order in which the item is displayed in the timeline, with a higher score appearing closer to the clock}

\item{recipients}{A list of users or groups that this item has been shared with}

\item{selfLink}{A URL that can be used to retrieve this item}

\item{sourceItemId}{Opaque string you can use to map a timeline item to data in your own service}

\item{speakableText}{The speakable version of the content of this item}

\item{speakableType}{A speakable description of the type of this item}

\item{text}{Text content of this item}

\item{title}{The title of this item}

\item{updated}{The time at which this item was last modified, formatted according to RFC 3339}
}
\value{
TimelineItem object
}
\description{
TimelineItem Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Each item in the user's timeline is represented as a TimelineItem JSON structure, described below.
}
\seealso{
Other TimelineItem functions: \code{\link{timeline.insert}},
  \code{\link{timeline.patch}},
  \code{\link{timeline.update}}
}
