% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_objects.R
\name{TimelineListResponse}
\alias{TimelineListResponse}
\title{TimelineListResponse Object}
\usage{
TimelineListResponse(items = NULL, nextPageToken = NULL)
}
\arguments{
\item{items}{Items in the timeline}

\item{nextPageToken}{The next page token}
}
\value{
TimelineListResponse object
}
\description{
TimelineListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of timeline items. This is the response from the server to GET requests on the timeline collection.
}
