% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_objects.R
\name{UserAction}
\alias{UserAction}
\title{UserAction Object}
\usage{
UserAction(payload = NULL, type = NULL)
}
\arguments{
\item{payload}{An optional payload for the action}

\item{type}{The type of action}
}
\value{
UserAction object
}
\description{
UserAction Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents an action taken by the user that triggered a notification.
}
