% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_objects.R
\name{UserData}
\alias{UserData}
\title{UserData Object}
\usage{
UserData(key = NULL, value = NULL)
}
\arguments{
\item{key}{No description}

\item{value}{No description}
}
\value{
UserData object
}
\description{
UserData Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

