% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_functions.R
\docType{package}
\name{mirror_googleAuthR}
\alias{mirror_googleAuthR}
\alias{mirror_googleAuthR-package}
\title{Google Mirror API
Interacts with Glass users via the timeline.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 20:02:07
filename: /Users/mark/dev/R/autoGoogleAPI/googlemirrorv1.auto/R/mirror_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/glass.location
\item https://www.googleapis.com/auth/glass.timeline
}
}
